rename create-post-config->create-request-config
