+++ client/src/components/NewTag/use-new-tag-mutation.ts
@@ -1,4 +1,4 @@
-import { createPostConfig } from "@lib/fetch/create-post-config";
+import { createPostConfig } from "@/lib/fetch/create-request-config";
 import { makeAuthorizedUrl } from "@lib/fetch/make-authorized-url";
 import { useMutation } from "@tanstack/react-query";
 import type { TagInput, TagWithIds } from "@type/server/tag.types";
