+++ client/src/lib/query/use-login-mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { Data } from "@type/query.types";
 import { User, UserLogin } from "@type/server/user.types";
-import { createPostConfig } from "../fetch/create-post-config";
+import { createPostConfig } from "../fetch/create-request-config";
 import { baseUrl } from "../fetch/fetch-constants";
 import { localUser } from "../user-storage";
