+++ client/src/lib/query/use-register-mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 import { Data } from "@type/query.types";
 import { NewUser, User } from "@type/server/user.types";
-import { createPostConfig } from "../fetch/create-post-config";
+import { createPostConfig } from "../fetch/create-request-config";
 import { baseUrl } from "../fetch/fetch-constants";
 
 async function postRegister(newUser: NewUser) {
